backend fixed, added flyway, unnecessary pom dependencies removed
