--- src/main/java/it/arcade/hospital/config/SecurityConfiguration.java
@@ -1,16 +1,0 @@
-package it.arcade.hospital.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.thymeleaf.extras.springsecurity4.dialect.SpringSecurityDialect;
-
-@Configuration
-public class SecurityConfiguration implements WebMvcConfigurer {
-
-    @Bean
-    public SpringSecurityDialect springSecurityDialect(){
-        return new SpringSecurityDialect();
-    }
-
-}
