--- src/main/java/it/arcade/hospital/department/Department.java
@@ -1,36 +1,0 @@
-package it.arcade.hospital.department;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-import it.arcade.hospital.user.doctor.Doctor;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-
-@Data
-@RequiredArgsConstructor
-@Entity
-@Table(name = "department")
-public class Department {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
-    int id;
-
-    @Column(name = "name")
-    String name;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "head_doctor")
-    Doctor headDoctor;
-}
