--- src/main/java/it/arcade/hospital/medicalrecord/MedicalRecord.java
@@ -1,65 +1,0 @@
-package it.arcade.hospital.medicalrecord;
-
-import java.util.Date;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-import org.hibernate.annotations.CreationTimestamp;
-
-import it.arcade.hospital.department.Department;
-import it.arcade.hospital.user.doctor.Doctor;
-import it.arcade.hospital.user.patient.Patient;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "medical_record")
-public class MedicalRecord {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
-    private int id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_patient")
-    private Patient patient;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_doctor")
-    private Doctor doctor;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_department")
-    private Department department;
-
-    @Column(name = "disease")
-    private String disease;
-
-    @Column(name = "description")
-    private String description;
-
-    @Column(name = "medicines")
-    private String medicines;
-
-    @CreationTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "entry_date")
-    private Date entryDate;
-}
