--- src/main/java/it/arcade/hospital/offer/Offer.java
@@ -1,37 +1,0 @@
-package it.arcade.hospital.offer;
-
-import java.math.BigDecimal;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.validation.constraints.NotNull;
-
-import it.arcade.hospital.department.Department;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Offer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    @NotNull
-    private String offerName;
-    @NotNull
-    private BigDecimal price;
-
-    @ManyToOne
-    @JoinColumn(name = "id_department")
-    private Department department;
-
-}
