--- src/main/java/it/arcade/hospital/reservation/Reservation.java
@@ -1,59 +1,0 @@
-package it.arcade.hospital.reservation;
-
-import java.util.Date;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-import org.hibernate.annotations.CreationTimestamp;
-
-import it.arcade.hospital.user.doctor.Doctor;
-import it.arcade.hospital.user.patient.Patient;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-
-@Data
-@RequiredArgsConstructor
-@Entity
-@Table(name = "reservation")
-public class Reservation {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
-    int id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_patient")
-    Patient patient;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "id_doctor")
-    Doctor doctor;
-
-    @CreationTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "start_date")
-    Date starDate; //todo replace with localdatetime
-
-    @CreationTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "end_date")
-    Date endDate;//todo replace with localdatetime
-
-    @CreationTimestamp
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "entry_date")
-    Date entryDate;//todo replace with localdatetime
-}
