--- src/main/java/it/arcade/hospital/role/RoleRepository.java
@@ -1,9 +1,0 @@
-package it.arcade.hospital.role;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-    Role findByRoleName(String roleName);
-}
