--- src/main/java/it/arcade/hospital/specialisation/Specialisation.java
@@ -1,27 +1,0 @@
-package it.arcade.hospital.specialisation;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-
-@Data
-@RequiredArgsConstructor
-@Entity
-@Table(name = "specialisation")
-public class Specialisation {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
-    int id;
-
-    @Column(name = "name")
-    String name;
-}
