+++ src/main/java/it/hospital/project/flyway/FlywayConfiguration.java
@@ -1,0 +1,18 @@
+package it.hospital.project.flyway;
+
+import org.springframework.boot.autoconfigure.flyway.FlywayMigrationStrategy;
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
+
+@Component
+public class FlywayConfiguration {
+
+    @Bean
+    public FlywayMigrationStrategy cleanMigrateStrategy() {
+
+        return flyway -> {
+            flyway.clean();
+            flyway.migrate();
+        };
+    }
+}
