+++ src/main/resources/application.properties
@@ -1,4 +1,7 @@
 server.port=9090
+
+
+#postgresql database
 spring.datasource.initialization-mode=always
 spring.datasource.initialize=true
 spring.datasource.password=root
@@ -6,7 +9,21 @@ spring.datasource.platform=postgres
 spring.datasource.url=jdbc:postgresql://localhost:5432/hospital_database
 spring.datasource.username=postgres
 spring.jpa.database=POSTGRESQL
-spring.jpa.generate-ddl=true
+spring.jpa.generate-ddl=false
 spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
-spring.jpa.show-sql=true
+
+
+
+#flyway properties
+spring.flyway.enabled=true
+spring.flyway.url=jdbc:postgresql://localhost:5432/hospital_database
+spring.flyway.user=postgres
+spring.flyway.password=root
+
+
+
+
+
+
+
+
