+++ src/main/resources/db/migration/afterMigrate.sql
@@ -1,0 +1,23 @@
+--User insertions
+insert into appuser (first_name, last_name , phoneNumber,
+  address, city, postcode, email, password) values('Jan',
+                                                   'Nowak',
+                                                   '+48500000000',
+                                                   'Pyszna 2/3',
+                                                   'Krakow',
+                                                   '30-126',
+                                                   'jnowak@example.com',
+                                                   '$2y$12$zKYvWsxOLrHHhc5wqn0yhu3/hPzjQHXDZBHewAKr6xaZl30bHSwNG ');
+
+
+--User insertion
+insert into appuser (first_name, last_name , phoneNumber,
+                  address, city, postcode, email, password) values('Marek',
+                                                                   'Gladysz',
+                                                                   '+48500000000',
+                                                                   'Ciepla 2/3',
+                                                                   'Makow',
+                                                                   '36-126',
+                                                                   'mgladysz@example.com',
+                                                                   '$2y$12$UFjY7Ex5jDomPCLNktoeG.sGaIJeX98TEiVy12mZ0D3adTtHjbOj6 ');
+
