+++ src/test/java/it/hospital/project/HospitalApplicationTests.java
@@ -1,4 +1,4 @@
-package it.arcade.hospital;
+package it.hospital.project;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
